refactor: extract insights cash flow chart
